+++ exporters/trace/stackdriver/src/main/java/io/opencensus/exporter/trace/stackdriver/StackdriverV2ExporterHandler.java
@@ -19,6 +19,7 @@ package io.opencensus.exporter.trace.stackdriver;
 import static com.google.api.client.util.Preconditions.checkNotNull;
 
 import com.google.api.gax.core.FixedCredentialsProvider;
+import com.google.api.gax.retrying.RetrySettings;
 import com.google.auth.Credentials;
 import com.google.cloud.trace.v2.TraceServiceClient;
 import com.google.cloud.trace.v2.TraceServiceSettings;
@@ -37,6 +38,7 @@ import com.google.devtools.cloudtrace.v2.TruncatableString;
 import com.google.protobuf.BoolValue;
 import com.google.protobuf.Int32Value;
 import com.google.rpc.Status;
+import io.opencensus.common.Duration;
 import io.opencensus.common.Function;
 import io.opencensus.common.Functions;
 import io.opencensus.common.OpenCensusLibraryInformation;
@@ -171,15 +173,42 @@ final class StackdriverV2ExporterHandler extends SpanExporter.Handler {
   static StackdriverV2ExporterHandler createWithCredentials(
       String projectId,
       Credentials credentials,
-      Map<String, io.opencensus.trace.AttributeValue> fixedAttributes)
+      Map<String, io.opencensus.trace.AttributeValue> fixedAttributes,
+      @javax.annotation.Nullable Duration deadline)
       throws IOException {
-    TraceServiceSettings traceServiceSettings =
+    TraceServiceSettings.Builder builder =
         TraceServiceSettings.newBuilder()
             .setCredentialsProvider(
-                FixedCredentialsProvider.create(checkNotNull(credentials, "credentials")))
-            .build();
+                FixedCredentialsProvider.create(checkNotNull(credentials, "credentials")));
+    if (deadline != null) {
+      // We only use the batchWriteSpans API in this exporter.
+      builder
+          .batchWriteSpansSettings()
+          .setSimpleTimeoutNoRetries(org.threeten.bp.Duration.ofMillis(deadline.toMillis()));
+    } else {
+      /*
+       * Default retry settings for Stackdriver Trace client is:
+       * settings =
+       *   RetrySettings.newBuilder()
+       *       .setInitialRetryDelay(Duration.ofMillis(100L))
+       *       .setRetryDelayMultiplier(1.2)
+       *       .setMaxRetryDelay(Duration.ofMillis(1000L))
+       *       .setInitialRpcTimeout(Duration.ofMillis(30000L))
+       *       .setRpcTimeoutMultiplier(1.5)
+       *       .setMaxRpcTimeout(Duration.ofMillis(60000L))
+       *       .setTotalTimeout(Duration.ofMillis(120000L))
+       *       .build();
+       *
+       * Override the default settings with settings that don't retry.
+       */
+      builder
+          .batchWriteSpansSettings()
+          .setRetryableCodes()
+          // RetrySettings.newBuilder().build() returns settings with no retries.
+          .setRetrySettings(RetrySettings.newBuilder().build());
+    }
     return new StackdriverV2ExporterHandler(
-        projectId, TraceServiceClient.create(traceServiceSettings), fixedAttributes);
+        projectId, TraceServiceClient.create(builder.build()), fixedAttributes);
   }
 
   @VisibleForTesting